Simplify running tests
There is no need to run the tests both by setup.py and nose: it just
runs the same tests twice. Also, generating coverage data is useless and
it will not be visible anywhere anyway. This allows us to get rid of one
build requires.

Also we should first run unittests and only then build the data for
dummy compose: if the tests fail, we get the info earlier.